% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outputs.R
\name{dc_clean_columns}
\alias{dc_clean_columns}
\title{Drop unnamed empty columns and make names unique}
\usage{
dc_clean_columns(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
cleaned data.frame
}
\description{
Drops columns that are BOTH unnamed (NA or "") and entirely NA.
Then makes remaining names unique.
}
\keyword{internal}
