% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{dc_list_experiments}
\alias{dc_list_experiments}
\title{List experiments for a site}
\usage{
dc_list_experiments(project, site, include_eq = TRUE)
}
\arguments{
\item{project}{Character. Project root path.}

\item{site}{Character. Site name.}

\item{include_eq}{Logical. Include "eq" if present.}
}
\value{
Character vector of experiment names.
}
\description{
Experiments are inferred primarily from subdirectories under:
\code{project/sites/{site}/outputs/}
}
