% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_project.R
\name{dc_list_sites}
\alias{dc_list_sites}
\title{List sites under a project}
\usage{
dc_list_sites(project, sites_dir = DDCENTVIZ_DEFAULT_SITES_DIR)
}
\arguments{
\item{project}{Character. Project root path.}

\item{sites_dir}{Character. Directory name under project root.}
}
\value{
Character vector of site folder names.
}
\description{
List sites under a project
}
