% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_project.R
\name{dc_project_root}
\alias{dc_project_root}
\title{Normalize and validate a DayCent project root}
\usage{
dc_project_root(
  path = NULL,
  sites_dir = DDCENTVIZ_DEFAULT_SITES_DIR,
  must_work = TRUE
)
}
\arguments{
\item{path}{Character. Project root path.}

\item{sites_dir}{Character. Directory name under project root (default "sites").}

\item{must_work}{Logical. If TRUE, error if path doesn't exist.}
}
\value{
Normalized project root path.
}
\description{
A DayCent project root is expected to contain a "sites/" directory.
If \code{path} is NULL, this will use \code{getwd()}.
}
