% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_water.R
\name{dc_read_layer_data}
\alias{dc_read_layer_data}
\title{Read DayCent layered output (vswc/wfps) into wide table with Layer_*}
\usage{
dc_read_layer_data(file_path)
}
\arguments{
\item{file_path}{Path to a DayCent layered output file (e.g., vswc.out, wfps.out).}
}
\value{
A data.frame/data.table with columns: time, dayofyr, Layer_*, date
}
\description{
Read DayCent layered output (vswc/wfps) into wide table with Layer_*
}
\keyword{internal}
