% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_project.R
\name{dc_site_path}
\alias{dc_site_path}
\title{Return the path to a site directory}
\usage{
dc_site_path(
  project,
  site,
  sites_dir = DDCENTVIZ_DEFAULT_SITES_DIR,
  must_work = TRUE
)
}
\arguments{
\item{project}{Character. Project root path.}

\item{site}{Character. Site name.}

\item{sites_dir}{Character. Directory name under project root.}

\item{must_work}{Logical. If TRUE, error if site dir doesn't exist.}
}
\value{
Normalized site path.
}
\description{
Return the path to a site directory
}
