% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_units.R
\name{g_m2_to_Mg_ha}
\alias{g_m2_to_Mg_ha}
\title{Convert SOC from g C / m2 to Mg C / ha}
\usage{
g_m2_to_Mg_ha(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
g/m2 -> Mg/ha:
1 g/m2 = 0.01 Mg/ha
}
