% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{infer_file_role}
\alias{infer_file_role}
\title{Infer a file "role" based on name + extension}
\usage{
infer_file_role(path)
}
\arguments{
\item{path}{Character. File path.}
}
\value{
Character role (one of DDCENTVIZ_ROLES).
}
\description{
Infer a file "role" based on name + extension
}
\keyword{internal}
