% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_carbon.R
\name{plot_carbon_pools}
\alias{plot_carbon_pools}
\title{Plot DayCent carbon pools (narrative)}
\usage{
plot_carbon_pools(data, scenario, start_year, end_year)
}
\arguments{
\item{data}{data.frame. Parsed dc_sip output.}

\item{scenario}{character. Label for plot titles.}

\item{start_year}{numeric. Start year (inclusive).}

\item{end_year}{numeric. End year (inclusive).}
}
\value{
This function returns plots for different categories of the C pool.
}
\description{
This function expects the DayCent dc_sip table as a data.frame and returns
a list of narrative ggplots.
}
\details{
This function was built to use outputs from a DayCent run to plot carbon pools over a specified period of time.
The output from this function is a list of seven plots:
\itemize{
\item Plot 1: plant carbon stocks
\item Plot 2: litter carbon stocks
\item Plot 3: soil organic carbon stocks
\item Plot 4: rapid turnover of soil carbon stocks
\item Plot 5: total system carbon stocks
\item Plot 6: turnover soil carbon stocks
\item Plot 7: relative turnover of soil carbon stocks
}
}
