% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{validate_manifest}
\alias{validate_manifest}
\title{Validate a manifest tibble/data.frame}
\usage{
validate_manifest(manifest, strict = FALSE)
}
\arguments{
\item{manifest}{data.frame.}

\item{strict}{Logical. If TRUE, stop on issues.}
}
\value{
manifest invisibly (possibly unchanged). Emits warnings/errors.
}
\description{
Validate a manifest tibble/data.frame
}
